Fix types, fix norm, fix NULL handling

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: pschwarz <pschwarz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/11 14:45:59 by pschwarz          #+#    #+#             */
-/*   Updated: 2022/11/03 13:51:08 by pschwarz         ###   ########.fr       */
+/*   Updated: 2022/11/04 18:12:40 by pschwarz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ size_t			ft_strlen(const char *s);
 void			*ft_memset(void *b, int c, int len);
 void			ft_bzero(void *s, size_t n);
 void			*ft_memcpy(void *dst, const void *src, size_t n);
-void			*ft_memmove(void *dst, const void *src, int len);
+void			*ft_memmove(void *dst, const void *src, size_t len);
 size_t			ft_strlcpy(char *dst, const char *src, size_t dstsize);
 size_t			ft_strlcat(char *dst, const char *src, size_t dstsize);
 int				ft_toupper(int c);
